requirements.txt
yfinance
requests
python-dotenv
fastapi
uvicorn
